# api/docs/swagger.yaml
basePath: /
definitions:
...
  model.Password:
    properties:
      current:
        type: string
      new:
        type: string
    type: object
...
  model.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      nick:
        type: string
      password:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: API responsible for CRUD and DevBook social network authentication.
  title: DevBook API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: ListUsersByFilter list users by a filter and returns the list of
        users found or an error.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Name or nick of user.
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: ListUsersByFilter list users by a filter.
      tags:
      - users
    post:
      consumes:
      - application/json
      description: CreateUser create a user and return the user created.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: ' '
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "422":
          description: 'Error: Unprocessable Entity'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: CreateUser create a user.
      tags:
      - users
  /users/{userId}:
    delete:
      consumes:
      - application/json
      description: DeleteUserById delete a user by the given ID and returns an error
        if unable to delete the user.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID.
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Success with no content
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "403":
          description: 'Error: Forbidden'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: DeleteUserById delete a user by the given ID.
      tags:
      - users
    get:
      consumes:
      - application/json
      description: FindUserById search for a user by the given ID and return a user
        or an error.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID.
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "404":
          description: 'Error: Not Found'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: FindUserById search for a user by the given ID.
      tags:
      - users
    put:
      consumes:
      - application/json
      description: UpdateUserById updates a user by the given ID and returns an error
        if unable to update the user.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID.
        in: path
        name: userId
        required: true
        type: integer
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "204":
          description: Success with no content
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "403":
          description: 'Error: Forbidden'
          schema:
            type: string
        "422":
          description: 'Error: Unprocessable Entity'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: UpdateUserById updates a user by the given ID.
      tags:
      - users
  /users/{userId}/follow:
    post:
      consumes:
      - application/json
      description: FollowUserById follows a user by the given ID and returns an error
        if unable to follow the user.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID.
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Success with no content
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "403":
          description: 'Error: Forbidden'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: FollowUserById follows a user by the given ID.
      tags:
      - users
  /users/{userId}/followers:
    get:
      consumes:
      - application/json
      description: ListFollowersByFollowedUserId lists a user's followers and returns
        a list of followers or an error.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID.
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: ListFollowersByFollowedUserId lists a user's followers.
      tags:
      - users
  /users/{userId}/list-followed:
    get:
      consumes:
      - application/json
      description: ListFollowedByFollowerId lists people the user follows and returns
        a list of people the user follows or an error.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID.
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: ListFollowedByFollowerId lists people the user follows.
      tags:
      - users
  /users/{userId}/unfollow:
    post:
      consumes:
      - application/json
      description: UnfollowUserById unfollow a user by the given ID and returns an
        error if unable to unfollow the user.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID.
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Success with no content
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "403":
          description: 'Error: Forbidden'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: UnfollowUserById unfollow a user by the given ID.
      tags:
      - users
  /users/{userId}/update-password:
    post:
      consumes:
      - application/json
      description: UpdatePasswordByUserId updates the user's password by the given
        ID and returns an error if unable to update the password.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID.
        in: path
        name: userId
        required: true
        type: integer
      - description: The current and new password of user.
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.Password'
      produces:
      - application/json
      responses:
        "204":
          description: Success with no content
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "403":
          description: 'Error: Forbidden'
          schema:
            type: string
        "422":
          description: 'Error: Unprocessable Entity'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: UpdatePasswordByUserId updates the user's password by the given ID.
      tags:
      - users
swagger: "2.0"